# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserInfo } from "./userInfo";

export interface LoginResponse {
  success: boolean;
  user: UserInfo;
}